Dropdown component refactoring
